backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
argon2-cffi==23.1.0
pyotp==2.9.0
python-jose==3.3.0
cryptography==42.0.5
pydantic-settings==2.6.1
email-validator==2.2.0
psycopg2-binary==2.9.10
